Updated Visual Studio build environment.

+++ src/widgets/wg_oscilloscope.cpp
@@ -485,7 +485,7 @@ namespace wg
 	  {
 		  m_pAAPix[m_iNextPixel] = Coord(x, y);
 		  m_pAACol[m_iNextPixel] = m_lineColor;
-		  m_pAACol[m_iNextPixel].a = (uint8_t)255*alpha;
+		  m_pAACol[m_iNextPixel].a = (uint8_t) (255*alpha);
 		  
 		  ++m_iNextPixel;
 	  }
@@ -548,8 +548,8 @@ namespace wg
 	   
 			xend = round(x0);
 			yend = y0 + gradient * (xend - x0);
-			xgap = rfpart(x0 + 0.5);
-			xpxl1 = xend;
+			xgap = rfpart(x0 + 0.5f);
+			xpxl1 = (int) xend;
 			ypxl1 = ipart(yend);
 			
 			if (steep) {
@@ -564,8 +564,8 @@ namespace wg
 			// Second end point
 			xend = round(x1);
 			yend = y1 + gradient * (xend - x1);
-			xgap = fpart(x1 + 0.5);
-			xpxl2 = xend;
+			xgap = fpart(x1 + 0.5f);
+			xpxl2 = (int) xend;
 			ypxl2 = ipart(yend);
 			
 			if (steep) {
